+++ src/hupper/__init__.py
@@ -2,6 +2,7 @@
 # flake8: noqa
 
 from .utils import is_watchdog_supported
+from .utils import is_watchman_supported
 
 from .reloader import (
     start_reloader,